core folder organized
