+++ core/hooks/useAuth.ts
@@ -1,5 +1,5 @@
+import { AuthContext } from "@core/contexts";
 import { useContext } from "react";
-import { AuthContext } from "../contexts/authContext";
 
 export const useAuth = () => {
   return useContext(AuthContext);
