+++ core/types/index.ts
@@ -1,6 +1,2 @@
-export * from "./template-props";
-export * from "./button-props";
-export * from "./auth-props";
-export * from "./roomCode-props";
-export * from "./databaseRoom-props";
-export * from "./questions-type";
+export * from "./props";
+export * from "./types";
