+++ core/types/props/databaseRoom-props.ts
