+++ core/types/props/index.ts
@@ -1,0 +1,5 @@
+export * from "./auth-props";
+export * from "./button-props";
+export * from "./databaseRoom-props";
+export * from "./roomCode-props";
+export * from "./template-props";
