+++ core/types/props/template-props.ts
