+++ core/types/types/index.ts
@@ -1,0 +1,2 @@
+export * from "./questions-type";
+export * from "./user-type";
