+++ pages/_app.tsx
@@ -1,5 +1,5 @@
-import { AuthProvider } from "@/contexts";
 import { Template } from "@components/layout";
+import { AuthProvider } from "@core/contexts";
 import type { AppProps } from "next/app";
 import "../styles/globals.css";
