+++ pages/rooms/[roomID]/index.tsx
@@ -2,8 +2,7 @@ import { app, db } from "@/firebase";
 import { logo } from "@/images";
 import { Question } from "@components/modules";
 import { Button, RoomCode } from "@components/widgets";
-import { useAuth } from "hooks/useAuth";
-import { useRoom } from "hooks/useRoom";
+import { useAuth, useRoom } from "@core/hooks/";
 import { NextPage } from "next";
 import Head from "next/head";
 import Image from "next/image";
